[student_notes] modernize this load test
* move EdxAppTasks out of the lms test and into helpers
* course enrollment support
* settings file support
* correct import style
* correct __init__.py
* add requrements.txt
* other adjustments which fix code making old assumptions (circa 2015)
