+++ loadtests/student_notes/__init__.py
@@ -1,0 +1,1 @@
+from locustfile import NotesLocust
