+++ loadtests/student_notes/requirements.txt
@@ -1,0 +1,3 @@
+# Packages needed to run this suite of loadtests
+
+-r ../../requirements/base.txt
